webapp-python/requirements.txt
boto3==1.26.16
botocore==1.29.16
click==8.1.3
Flask==2.2.2
flask-dynamo==0.1.2
importlib-metadata==5.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
MarkupSafe==2.1.1
python-dateutil==2.8.2
s3transfer==0.6.0
six==1.16.0
urllib3==1.26.13
Werkzeug==2.2.2
zipp==3.10.0
pytz==2022.6
python-dotenv==0.21.0